third_party/blink/web_tests/external/wpt/svg/linking/scripted/a.rel-getter-01-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's rel getter
  assert_equals: expected (string) "noreferrer" but got (undefined) undefined
Harness: the test ran to completion.
